+++ Magic/CHANGELOG.md
@@ -5,6 +5,7 @@
  - Fix loading on old versions of Spigot
  - Fix errors using `/mgive someitem someamount` without a player target
  - Some improvements to Wolf House
+ - Fix issues with SQLLite on older Spigot versions
 
 # 7.10.5